Python and shell scripts for thumbnail management.
dirtothumbs.sh iterates over files in the cwd, saving thumbnails to thumbs.
squarethumb.sh can be passed a list of files, and does the same.
dirdiff.py compares two directories: this allows me to see where a thumbnail is
missing, due to a processing errorPython and shell scripts for thumbnail
management.

dirtothumbs.sh iterates over files in the cwd, saving thumbnails to thumbs.
squarethumb.sh can be passed a list of files, and does the same.
dirdiff.py compares two directories: this allows me to see where a thumbnail is
missing, due to a processing error..